requirements.txt
langchain-core>=0.1.33
langchain-community>=0.0.20
langchain-openai>=0.0.8
langchain-text-splitters>=0.0.1
langchain-pinecone>=0.0.2
pinecone-client>=3.0.0
openai>=1.12.0
python-dotenv>=1.0.0
streamlit>=1.29.0
pypdf>=3.17.0
tqdm>=4.66.1
markdown>=3.5.2
weasyprint>=57.0